+++ customs/brc20/src/lib.rs
@@ -22,6 +22,8 @@ pub mod constants {
     pub const FETCH_HUB_DIRECTIVE_NAME: &str = "FETCH_HUB_DIRECTIVE";
     pub const FINALIZE_GENERATE_TICKET_NAME: &str = "FINALIZE_GENERATE_TICKET_NAME";
     pub const FINALIZE_GENERATE_TICKET_INTERVAL: u64 = 600;
+    pub const FINALIZE_TO_BITCOIN_TICKET_NAME: &str = "FINALIZE_TO_BITCOIN_TICKET_NAME";
+    pub const FINALIZE_TO_BITCOIN_TICKET_INTERVAL: u64 = 600;
     pub const BATCH_QUERY_LIMIT: u64 = 20;
     pub const PROD_KEY: &str = "key_1";
     pub const SEC_NANOS: u64 = 1_000_000_000;
